feat: class btn 해결 못함,, filter modal창으로 변경
