+++ src/styles/Modal.module.css
@@ -7,6 +7,7 @@
   z-index: 1;
   width: 100%;
   position: fixed;
+  color: var(--color-white);
 }
 
 .paddingContainer {
@@ -37,6 +38,14 @@
   display: flex;
   align-items: center;
   justify-content: right;
+  width: 100%;
+  height: 55px;
+}
+
+.filterCloseBox {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
   width: 100%;
   height: 55px;
 }
